requirements.txt
WTForms==2.3.3
Flask==1.1.2
passlib==1.7.2
Flask_Login==0.5.0
Flask_WTF==0.14.3
mysql_connector_repackaged==0.3.1
gunicorn==20.0.4
mysql-connector-python ==8.0.22